requirements.txt
absl-py==0.9.0
bz2file==0.98
flask==1.1.2
future==0.18.2
gevent==20.6.2
gin-config==0.3.0
google-api-python-client==1.9.3
gunicorn==20.0.4
gym==0.17.2
h5py==2.10.0
kfac==0.2.0
mesh-tensorflow==0.1.13
numpy==1.18.5
oauth2client==4.1.3
opencv-python==4.2.0.34
Pillow==7.1.2
pypng==0.0.20
requests==2.23.0
scipy==1.4.1
six==1.15.0
sympy==1.6.1
tensorflow-addons==0.6.0
tensorflow-datasets==3.1.0
tensorflow-gan==2.0.0
tensorflow-probability==0.7.0
tf_slim==1.1.0
tqdm==4.46.1